+++ crates/noseyparker/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod blob;
 pub mod blob_id;
 pub mod blob_id_set;
+pub mod blob_metadata;
 pub mod bstring_escape;
 pub mod datastore;
 pub mod defaults;
@@ -13,8 +14,7 @@ pub mod location;
 pub mod match_type;
 pub mod matcher;
 pub mod matcher_stats;
-#[cfg(feature = "content_guesser")]
-pub use noseyparker_content_guesser as content_guesser;
+pub use content_guesser;
 pub mod progress;
 pub mod provenance;
 #[cfg(feature = "rule_profiling")]